64 bits port
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_0_0_dev@609 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/PlanarityTest.cpp
@@ -57,10 +57,10 @@ list<edge> PlanarityTest::getObstructionsEdges(Graph *graph) {
 //=================================================================
 bool PlanarityTest::compute(Graph *graph) { 
 
-  if (resultsBuffer.find((unsigned int)graph)!=resultsBuffer.end()) 
-    return resultsBuffer[(unsigned int)graph];
+  if (resultsBuffer.find((unsigned long)graph)!=resultsBuffer.end()) 
+    return resultsBuffer[(unsigned long)graph];
   else if(graph->numberOfNodes()==0){
-    resultsBuffer[(unsigned int)graph] = true;
+    resultsBuffer[(unsigned long)graph] = true;
     return true;
   }
   vector<edge> addedEdges;
